refactor: Criação do pacote role para os DTOs
